refactor defaults per-OS
