+++ pkg/config/defaults_darwin.go
@@ -11,6 +11,7 @@ import (
 	"github.com/docker/go-units"
 	"github.com/xorcare/pointer"
 
+	"github.com/runfinch/finch/pkg/command"
 	"github.com/runfinch/finch/pkg/fmemory"
 )
 
@@ -61,4 +62,24 @@ func cpuDefault(cfg *Finch, deps LoadSystemDeps) {
 			cfg.CPUs = pointer.Int(fallbackCPUs)
 		}
 	}
+}
+
+// applyDefaults sets default configuration options if they are not already set.
+func applyDefaults(
+	cfg *Finch,
+	deps LoadSystemDeps,
+	mem fmemory.Memory,
+	ecc command.Creator,
+) *Finch {
+	cpuDefault(cfg, deps)
+	memoryDefault(cfg, mem)
+	supportsVz := false
+	vz, err := SupportsVirtualizationFramework(ecc)
+	if err == nil && vz {
+		supportsVz = true
+	}
+	vmDefault(cfg, supportsVz)
+	rosettaDefault(cfg, supportsVz)
+
+	return cfg
 }
